requirements_runtime.txt
# Runtime dependencies placeholder for road_mome_demo_v1
# TODO: Add only validated runtime packages required by the demo software.
absl-py==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asttokens==3.0.1
attrs==25.4.0
blinker==1.9.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.3.1
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
colorama==0.4.6
comm==0.2.3
ConfigArgParse==1.7.1
contourpy==1.3.2
cycler==0.12.1
dash==3.3.0
decorator==5.2.1
docker-pycreds==0.4.0
easydict==1.9
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.128.0
fastjsonschema==2.21.2
filelock==3.20.0
Flask==3.1.2
fonttools==4.61.1
fsspec==2025.12.0
ftfy==6.0.1
gitdb==4.0.12
GitPython==3.1.45
grpcio==1.76.0
h11==0.16.0
h5py==3.6.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.3
idna==3.4
importlib_metadata==8.7.1
ipython==8.37.0
ipywidgets==8.1.8
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_core==5.9.1
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
line_profiler==5.0.0
lmdb==1.3.0
Markdown==3.10.1
MarkupSafe==2.1.5
matplotlib==3.10.8
matplotlib-inline==0.2.1
mpmath==1.3.0
narwhals==2.14.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
open-clip-torch==2.24.0
open3d==0.19.0
opencv-python==4.11.0.86
packaging==25.0
parso==0.8.5
pathtools==0.1.2
pillow==12.0.0
platformdirs==4.5.1
plotly==6.5.0
promise==2.3
prompt_toolkit==3.0.52
protobuf==3.20.3
psutil==7.2.0
pure_eval==0.2.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
pyyaml_env_tag==0.1
referencing==0.37.0
regex==2022.10.31
requests==2.28.1
retrying==1.4.2
rpds-py==0.30.0
scikit-learn==1.7.2
scipy==1.15.3
sentencepiece==0.2.1
sentry-sdk==2.48.0
setproctitle==1.3.7
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
starlette==0.50.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==1.1.0
threadpoolctl==3.6.0
timm==0.4.12
tomli==2.3.0
torch==2.1.2+cu118
torchaudio==2.1.2+cu118
torchvision==0.16.2+cu118
tqdm==4.67.1
traitlets==5.14.3
typer-slim==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==1.26.13
uvicorn==0.40.0
wandb==0.13.3
wcwidth==0.2.14
Werkzeug==3.1.4
widgetsnbextension==4.0.15
zipp==3.23.0
